requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
psycopg[binary]       
pydantic-settings
python-dotenv
